[#278] Moved ChannelOptions to the Ably::Models module. Updated failed tests. Implemented channel_options as ChannelOptions object instead of Hash.
