+++ lib/ably/realtime/presence/members_map.rb
@@ -207,7 +207,7 @@ module Ably::Realtime
 
       def setup_event_handlers
         presence.__incoming_msgbus__.subscribe(:presence, :sync) do |presence_message|
-          presence_message.decode(client.encoders, channel.options) do |encode_error, error_message|
+          presence_message.decode(client.encoders, channel.options.params) do |encode_error, error_message|
             client.logger.error error_message
           end
           update_members_and_emit_events presence_message
