+++ lib/ably/rest/presence.rb
@@ -89,7 +89,7 @@ module Ably
       end
 
       def decode_message(presence_message)
-        presence_message.decode client.encoders, channel.options
+        presence_message.decode client.encoders, channel.options.params
       rescue Ably::Exceptions::CipherError, Ably::Exceptions::EncoderError => e
         client.logger.error { "Decoding Error on presence channel '#{channel.name}', presence message client_id '#{presence_message.client_id}'. #{e.class.name}: #{e.message}" }
       end
